Add interface_type parameter
This parameter allows to specify the interface type. Currently only
'system', 'internal' and 'tap' are supported. Any types like tunnels
that need further configuration are not (yet) supported.

The parameter is only supported on providers that have the
interface_type feature. This feature is currently only supported on the
ovs provider.

Change-Id: Iac65696c0ff714395f81529fc091d44fed92c173